tickets/DM-33509B: fix test suite for kubernetes_asyncio

+++ tests/conftest.py
@@ -1,4 +1,5 @@
 """pytest fixtures for kubespawner"""
+import asyncio
 import base64
 import inspect
 import io
@@ -7,12 +8,13 @@ import os
 import sys
 import tarfile
 import time
-from distutils.version import LooseVersion as V
+from packaging.version import Version as V
 from functools import partial
-from threading import Thread
+from threading import Thread, Event
 
 import kubernetes_asyncio
 import pytest
+import pytest_asyncio
 from jupyterhub.app import JupyterHub
 from jupyterhub.objects import Hub
 from kubernetes_asyncio.client import V1ConfigMap
@@ -25,16 +27,29 @@ from kubernetes_asyncio.client import V1ServicePort
 from kubernetes_asyncio.client import V1ServiceSpec
 from kubernetes_asyncio.client.rest import ApiException
 from kubernetes_asyncio.config import load_kube_config
-#from kubernetes.stream import stream
-#from kubernetes_asyncio.stream import stream
 from kubernetes_asyncio.watch import Watch
 from traitlets.config import Config
 
 from kubespawner.clients import shared_client
 
+# Needed for the streaming stuff
+from kubernetes.stream import stream as sync_stream
+from kubernetes.config import load_kube_config as sync_load_kube_config
+from kubernetes.client import CoreV1Api as sync_CoreV1Api
+
 here = os.path.abspath(os.path.dirname(__file__))
 jupyterhub_config_py = os.path.join(here, "jupyterhub_config.py")
 
+# We do these to set up the synchronous client, needed for executing
+# python inside pods.
+sync_load_kube_config()
+sync_corev1api = sync_CoreV1Api()
+
+@pytest.fixture(scope="session")
+def event_loop():
+    loop=asyncio.get_event_loop()
+    yield loop
+    loop.close()
 
 @pytest.fixture(autouse=True)
 def traitlets_logging():
@@ -52,7 +67,6 @@ def traitlets_logging():
 def kube_ns():
     """Fixture for the kubernetes namespace"""
     return os.environ.get("KUBESPAWNER_TEST_NAMESPACE") or "kubespawner-test"
-
 
 @pytest.fixture
 def config(kube_ns):
@@ -102,20 +116,18 @@ def ssl_app(tmpdir_factory, kube_ns):
     return app
 
 
-def watch_logs(kube_client, pod_info):
+async def watch_logs(kube_client, pod_info):
     """Stream a single pod's logs
 
-    pod logs are streamed directly to sys.stderr,
+    pod logs are streamed directly to sys.stdout,
     so that pytest capture can deal with it.
-
-    Blocking, should be run in a thread.
 
     Called for each new pod from watch_kubernetes
     """
     watch = Watch()
     while True:
         try:
-            for event in watch.stream(
+            async for event in watch.stream(
                 func=kube_client.read_namespaced_pod_log,
                 namespace=pod_info.namespace,
                 name=pod_info.name,
@@ -125,53 +137,72 @@ def watch_logs(kube_client, pod_info):
             if e.status == 400:
                 # 400 can occur if the container is not yet ready
                 # wait and retry
-                time.sleep(1)
+                await asyncio.sleep(1)
                 continue
             elif e.status == 404:
                 # pod is gone, we are done
                 return
             else:
-                # unexpeced error
+                # unexpected error
                 print(f"Error watching logs for {pod_info.name}: {e}", file=sys.stderr)
                 raise
         else:
             break
 
 
-def watch_kubernetes(kube_client, kube_ns):
+async def watch_kubernetes(kube_client, kube_ns):
     """Stream kubernetes events to stdout
 
     so that pytest io capturing can include k8s events and logs
 
     All events are streamed to stdout
 
-    When a new pod is started, spawn an additional thread to watch its logs
+    When a new pod is started, spawn an additional task to watch its logs
     """
-    log_threads = {}
+
     watch = Watch()
-    for event in watch.stream(
-        func=kube_client.list_namespaced_event,
-        namespace=kube_ns,
-    ):
+    watch_task = {}
 
-        resource = event['object']
-        obj = resource.involved_object
-        print(f"k8s event ({event['type']} {obj.kind}/{obj.name}): {resource.message}")
-
-        # new pod appeared, start streaming its logs
-        if (
-            obj.kind == "Pod"
-            and event["type"] == "ADDED"
-            and obj.name not in log_threads
+    try:
+        async for event in watch.stream(
+            func=kube_client.list_namespaced_event,
+            namespace=kube_ns,
         ):
-            log_threads[obj.name] = t = Thread(
-                target=watch_logs, args=(kube_client, obj), daemon=True
-            )
-            t.start()
+
+            resource = event['object']
+            obj = resource.involved_object
+            print(f"k8s event ({event['type']} {obj.kind}/{obj.name}): {resource.message}")
+
+            # new pod appeared, start streaming its logs
+            if (
+                obj.kind == "Pod"
+                and event["type"] == "ADDED"
+                and obj.name not in watch_task
+            ):
+                watch_task[obj.name] = asyncio.create_task(
+                    watch_logs(
+                        kube_client,
+                        obj,
+                    ),
+                )
+
+    except asyncio.CancelledError as exc:
+        # kube_client cleanup cancelled us.  In turn, we should cancel
+        # the individual watch tasks.
+        await stop_signal.get()
+        for t in watch_task:
+            if watch_task[t] and not watch_task[t].done():
+                try:
+                    watch_task[t].cancel()
+                except asyncio.CancelledError:
+                    # Swallow these; they are what we expect.
+                    pass
+        # And re-raise so kube_client can finish cleanup
+        raise exc
 
 
-@pytest.fixture(scope="session")
-def kube_client(request, kube_ns):
+@pytest_asyncio.fixture(scope="session")
+async def kube_client(request, kube_ns):
     """fixture for the Kubernetes client object.
 
     skips test that require kubernetes if kubernetes cannot be contacted
@@ -180,29 +211,38 @@ def kube_client(request, kube_ns):
     - Hooks up kubernetes events and logs to pytest capture
     - Cleans up kubernetes namespace on exit
     """
-    load_kube_config()
+    await load_kube_config()
     client = shared_client("CoreV1Api")
+    stop_signal=asyncio.Queue()
     try:
-        namespaces = client.list_namespace(_request_timeout=3)
+        namespaces = await client.list_namespace(_request_timeout=3)
     except Exception as e:
         pytest.skip("Kubernetes not found: %s" % e)
 
     if not any(ns.metadata.name == kube_ns for ns in namespaces.items):
         print("Creating namespace %s" % kube_ns)
-        client.create_namespace(V1Namespace(metadata=dict(name=kube_ns)))
+        await client.create_namespace(V1Namespace(metadata=dict(name=kube_ns)))
     else:
         print("Using existing namespace %s" % kube_ns)
 
     # begin streaming all logs and events in our test namespace
-    t = Thread(target=watch_kubernetes, args=(client, kube_ns), daemon=True)
-    t.start()
+    t = asyncio.create_task(watch_kubernetes(client, kube_ns))
 
-    # delete the test namespace when we finish
-    def cleanup_namespace():
-        client.delete_namespace(kube_ns, body={}, grace_period_seconds=0)
-        for i in range(3):
+    yield client
+
+    # Clean up at close by sending a cancel to watch_kubernetes and letting
+    # it handle the signal, cancel the tasks *it* started, and then raising
+    # it back to us. 
+    try:
+        t.cancel()
+    except asyncio.CancelledError:
+        pass
+    # allow opting out of namespace cleanup, for post-mortem debugging
+    if not os.environ.get("KUBESPAWNER_DEBUG_NAMESPACE"):
+        await client.delete_namespace(kube_ns, body={}, grace_period_seconds=0)
+        for i in range(20):  # Usually finishes a good deal faster
             try:
-                ns = client.read_namespace(kube_ns)
+                ns = await client.read_namespace(kube_ns)
             except ApiException as e:
                 if e.status == 404:
                     return
@@ -210,19 +250,14 @@ def kube_client(request, kube_ns):
                     raise
             else:
                 print("waiting for %s to delete" % kube_ns)
-                time.sleep(1)
-
-    # allow opting out of namespace cleanup, for post-mortem debugging
-    if not os.environ.get("KUBESPAWNER_DEBUG_NAMESPACE"):
-        request.addfinalizer(cleanup_namespace)
-    return client
+                await asyncio.sleep(1)
 
 
-def wait_for_pod(kube_client, kube_ns, pod_name, timeout=90):
+async def wait_for_pod(kube_client, kube_ns, pod_name, timeout=90):
     """Wait for a pod to be ready"""
     conditions = {}
     for i in range(int(timeout)):
-        pod = kube_client.read_namespaced_pod(namespace=kube_ns, name=pod_name)
+        pod = await kube_client.read_namespaced_pod(namespace=kube_ns, name=pod_name)
         for condition in pod.status.conditions or []:
             conditions[condition.type] = condition.status
 
@@ -230,7 +265,7 @@ def wait_for_pod(kube_client, kube_ns, pod_name, timeout=90):
             print(
                 f"Waiting for pod {kube_ns}/{pod_name}; current status: {pod.status.phase}; {conditions}"
             )
-            time.sleep(1)
+            await asyncio.sleep(1)
         else:
             break
 
@@ -239,7 +274,7 @@ def wait_for_pod(kube_client, kube_ns, pod_name, timeout=90):
     return pod
 
 
-def ensure_not_exists(kube_client, kube_ns, name, resource_type, timeout=30):
+async def ensure_not_exists(kube_client, kube_ns, name, resource_type, timeout=30):
     """Ensure an object doesn't exist
 
     Request deletion and wait for it to be gone
@@ -247,7 +282,7 @@ def ensure_not_exists(kube_client, kube_ns, name, resource_type, timeout=30):
     delete = getattr(kube_client, "delete_namespaced_{}".format(resource_type))
     read = getattr(kube_client, "read_namespaced_{}".format(resource_type))
     try:
-        delete(namespace=kube_ns, name=name)
+        await delete(namespace=kube_ns, name=name)
     except ApiException as e:
         if e.status != 404:
             raise
@@ -255,7 +290,7 @@ def ensure_not_exists(kube_client, kube_ns, name, resource_type, timeout=30):
     while True:
         # wait for delete
         try:
-            read(namespace=kube_ns, name=name)
+            await read(namespace=kube_ns, name=name)
         except ApiException as e:
             if e.status == 404:
                 # deleted
@@ -264,10 +299,10 @@ def ensure_not_exists(kube_client, kube_ns, name, resource_type, timeout=30):
                 raise
         else:
             print("waiting for {}/{} to delete".format(resource_type, name))
-            time.sleep(1)
+            await asyncio.sleep(1)
 
 
-def create_resource(kube_client, kube_ns, resource_type, manifest, delete_first=True):
+async def create_resource(kube_client, kube_ns, resource_type, manifest, delete_first=True):
     """Create a kubernetes resource
 
     handling 409 errors and others that can occur due to rapid startup
@@ -275,13 +310,13 @@ def create_resource(kube_client, kube_ns, resource_type, manifest, delete_first=
     """
     name = manifest.metadata["name"]
     if delete_first:
-        ensure_not_exists(kube_client, kube_ns, name, resource_type)
+        await ensure_not_exists(kube_client, kube_ns, name, resource_type)
     print(f"Creating {resource_type} {name}")
     create = getattr(kube_client, f"create_namespaced_{resource_type}")
     error = None
     for i in range(10):
         try:
-            create(
+            await create(
                 body=manifest,
                 namespace=kube_ns,
             )
@@ -291,14 +326,14 @@ def create_resource(kube_client, kube_ns, resource_type, manifest, delete_first=
             error = e
             # need to retry since this can fail if run too soon after namespace creation
             print(e, file=sys.stderr)
-            time.sleep(int(e.headers.get("Retry-After", 1)))
+            await asyncio.sleep(int(e.headers.get("Retry-After", 1)))
         else:
             break
     else:
         raise error
 
 
-def create_hub_pod(kube_client, kube_ns, pod_name="hub", ssl=False):
+async def create_hub_pod(kube_client, kube_ns, pod_name="hub", ssl=False):
     config_map_name = pod_name + "-config"
     secret_name = pod_name + "-secret"
     with open(jupyterhub_config_py) as f:
@@ -308,7 +343,7 @@ def create_hub_pod(kube_client, kube_ns, pod_name="hub", ssl=False):
         metadata={"name": config_map_name}, data={"jupyterhub_config.py": config}
     )
 
-    config_map = create_resource(
+    config_map = await create_resource(
         kube_client,
         kube_ns,
         "config_map",
@@ -361,14 +396,14 @@ def create_hub_pod(kube_client, kube_ns, pod_name="hub", ssl=False):
             ],
         ),
     )
-    pod = create_resource(kube_client, kube_ns, "pod", pod_manifest)
-    return wait_for_pod(kube_client, kube_ns, pod_name)
+    pod = await create_resource(kube_client, kube_ns, "pod", pod_manifest)
+    return await wait_for_pod(kube_client, kube_ns, pod_name)
 
 
-@pytest.fixture(scope="session")
-def hub_pod(kube_client, kube_ns):
+@pytest_asyncio.fixture(scope="session")
+async def hub_pod(kube_client, kube_ns):
     """Create and return a pod running jupyterhub"""
-    return create_hub_pod(kube_client, kube_ns)
+    return await create_hub_pod(kube_client, kube_ns)
 
 
 @pytest.fixture
@@ -380,8 +415,8 @@ def hub(hub_pod):
     return Hub(ip=hub_pod.status.pod_ip, port=8081)
 
 
-@pytest.fixture(scope="session")
-def hub_pod_ssl(kube_client, kube_ns, ssl_app):
+@pytest_asyncio.fixture(scope="session")
+async def hub_pod_ssl(kube_client, kube_ns, ssl_app):
     """Start a hub pod with internal_ssl enabled"""
     # load ssl dir to tarfile
     buf = io.BytesIO()
@@ -394,7 +429,7 @@ def hub_pod_ssl(kube_client, kube_ns, ssl_app):
     secret_manifest = V1Secret(
         metadata={"name": secret_name}, data={"internal-ssl.tar": b64_certs}
     )
-    create_resource(kube_client, kube_ns, "secret", secret_manifest)
+    await create_resource(kube_client, kube_ns, "secret", secret_manifest)
 
     name = "hub-ssl"
 
@@ -407,9 +442,9 @@ def hub_pod_ssl(kube_client, kube_ns, ssl_app):
         ),
     )
 
-    create_resource(kube_client, kube_ns, "service", service_manifest)
+    await create_resource(kube_client, kube_ns, "service", service_manifest)
 
-    return create_hub_pod(
+    return await create_hub_pod(
         kube_client,
         kube_ns,
         pod_name=name,
@@ -444,7 +479,7 @@ class ExecError(Exception):
         )
 
 
-def _exec_python_in_pod(kube_client, kube_ns, pod_name, code, kwargs=None, _retries=0):
+async def _exec_python_in_pod(kube_client, kube_ns, pod_name, code, kwargs=None, _retries=0):
     """Run simple Python code in a pod
 
     code can be a str of code, or a 'simple' Python function,
@@ -452,11 +487,11 @@ def _exec_python_in_pod(kube_client, kube_ns, pod_name, code, kwargs=None, _retr
 
     kwargs are passed to the function, if it is given.
     """
-    if V(kubernetes.__version__) < V("11"):
+    if V(kubernetes_asyncio.__version__) < V("11"):
         pytest.skip(
             f"exec tests require kubernetes >= 11, got {kubernetes.__version__}"
         )
-    pod = wait_for_pod(kube_client, kube_ns, pod_name)
+    pod = await wait_for_pod(kube_client, kube_ns, pod_name)
     original_code = code
     if not isinstance(code, str):
         # allow simple self-contained (no globals or args) functions
@@ -480,8 +515,14 @@ def _exec_python_in_pod(kube_client, kube_ns, pod_name, code, kwargs=None, _retr
     print("Running {} in {}".format(code, pod_name))
     # need to create ws client to get returncode,
     # see https://github.com/kubernetes-client/python/issues/812
-    client = stream(
-        kube_client.connect_get_namespaced_pod_exec,
+    #
+    # That's why we are using the synchronous Kubernetes client here
+    # and why we imported them in the first place: kubernetes_asyncio
+    # does not yet support multichannel ws clients, which are needed
+    # to get the return code.
+    # cf https://github.com/tomplus/kubernetes_asyncio/issues/12
+    client = sync_stream(
+        sync_corev1api.connect_get_namespaced_pod_exec,
         pod_name,
         namespace=kube_ns,
         command=exec_command,
@@ -504,8 +545,8 @@ def _exec_python_in_pod(kube_client, kube_ns, pod_name, code, kwargs=None, _retr
             raise ExecError(exit_code=returncode, message=stderr, command=code)
         else:
             # retry
-            time.sleep(1)
-            return _exec_python_in_pod(
+            await asyncio.sleep(1)
+            return await _exec_python_in_pod(
                 kube_client,
                 kube_ns,
                 pod_name,
@@ -525,8 +566,8 @@ def exec_python_pod(kube_client, kube_ns):
     return partial(_exec_python_in_pod, kube_client, kube_ns)
 
 
-@pytest.fixture(scope="session")
-def exec_python(kube_ns, kube_client):
+@pytest_asyncio.fixture(scope="session")
+async def exec_python(kube_client, kube_ns):
     """Return a callable to execute Python code in a pod in the test namespace
 
     This fixture creates a dedicated pod for executing commands
@@ -551,6 +592,6 @@ def exec_python(kube_ns, kube_client):
             termination_grace_period_seconds=0,
         ),
     )
-    pod = create_resource(kube_client, kube_ns, "pod", pod_manifest)
+    pod = await create_resource(kube_client, kube_ns, "pod", pod_manifest)
 
     yield partial(_exec_python_in_pod, kube_client, kube_ns, pod_name)
